test/test/kharitonov/day5/file/FileTextReaderTest.java
package test.kharitonov.day5.file;

import by.kharitonov.day5.exception.TextProcessingException;
import by.kharitonov.day5.file.FileTextReader;
import by.kharitonov.day5.type.TextProcessingResults;
import org.testng.annotations.DataProvider;
import org.testng.annotations.Parameters;
import org.testng.annotations.Test;

import static org.testng.Assert.assertEquals;
import static org.testng.Assert.fail;

public class FileTextReaderTest {
    private final FileTextReader fileTextReader = new FileTextReader();

    @Test(groups = {"fileChar", "fileRegEx", "fileString"})
    public void testRead() {
        String fileName = "resources\\SourceText.txt";
        String expectedString = TextProcessingResults.SOURCE_TEXT;
        try {
            String actual = fileTextReader.read(fileName);
            assertEquals(actual, expectedString);
        } catch (TextProcessingException e) {
            fail();
        }
    }

    @DataProvider(name = "dataReadException")
    @Test
    public Object[][] dataReadEXception() {
        return new Object[][]{
                {null},
                {"WrongFile.txt"}
        };
    }

    @Parameters("fileName")
    @Test(expectedExceptions = TextProcessingException.class,
            dataProvider = "dataReadException",
            groups = {"fileChar", "fileRegEx", "fileString"})
    public void testReadException(String fileName)
            throws TextProcessingException {
        fileTextReader.read(fileName);
    }
}